assimp

+++ engine/Runtime/UI/Image.cpp
@@ -44,7 +44,7 @@ namespace DivineBrush::UI {
                 {{texture2d->GetWidth(), texture2d->GetHeight(), 0.0f}, {1.0f, 1.0f, 1.0f, 1.0f}, {1.f, 1.f}},
                 {{0.f,                   texture2d->GetHeight(), 0.0f}, {1.0f, 1.0f, 1.0f, 1.0f}, {0.f, 1.f}}
         };
-        std::vector<unsigned short> index_vector = {
+        std::vector<unsigned int> index_vector = {
                 0, 1, 2,
                 0, 2, 3
         };
